backend/requirements.txt
Django>=5.1,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
qrcode[pil]>=7.4
Pillow>=10.0
